+++ R/colour-manip.R
@@ -53,7 +53,7 @@ muted <- function(colour, l = 30, c = 70) col2hcl(colour, l = l, c = c)
 alpha <- function(colour, alpha = NA) {
   if (length(colour) != length(alpha)) {
     if (length(colour) > 1 && length(alpha) > 1) {
-      stop("Only one of colour and alpha can be vectorised")
+      cli::cli_abort("Only one of {.arg colour} and {.arg alpha} can be vectorised")
     }
 
     if (length(colour) > 1) {