+++ R/pal-brewer.R
@@ -49,7 +49,7 @@ brewer_pal <- pal_brewer
 pal_name <- function(palette, type) {
   if (is.character(palette)) {
     if (!palette %in% unlist(brewer)) {
-      warning("Unknown palette ", palette)
+      cli::cli_warn("Unknown palette: {.val {palette}}")
       palette <- "Greens"
     }
     return(palette)