+++ R/trans.R
@@ -1,39 +1,50 @@
 #' Create a new transformation object
 #'
 #' A transformation encapsulates a transformation and its inverse, as well
-#' as the information needed to create pleasing breaks and labels. The breaks
-#' function is applied on the transformed range of the range, and it's
-#' expected that the labels function will perform some kind of inverse
-#' transformation on these breaks to give them labels that are meaningful on
-#' the original scale.
+#' as the information needed to create pleasing breaks and labels. The `breaks()`
+#' function is applied on the un-transformed range of the data, and the
+#' `format()` function takes the output of the `breaks()` function and returns
+#' well-formatted labels. Transformations may also include the derivatives of the
+#' transformation and its inverse, but are not required to.
 #'
 #' @param name transformation name
 #' @param transform function, or name of function, that performs the
 #'   transformation
 #' @param inverse function, or name of function, that performs the
 #'   inverse of the transformation
+#' @param d_transform Optional function, or name of function, that gives the
+#'   derivative of the transformation. May be `NULL`.
+#' @param d_inverse Optional function, or name of function, that gives the
+#'   derivative of the inverse of the transformation. May be `NULL`.
 #' @param breaks default breaks function for this transformation. The breaks
-#'   function is applied to the raw data.
+#'   function is applied to the un-transformed data.
 #' @param minor_breaks default minor breaks function for this transformation.
 #' @param format default format for this transformation. The format is applied
-#'   to breaks generated to the raw data.
-#' @param domain domain, as numeric vector of length 2, over which
-#'   transformation is valued
+#'   to breaks generated on the un-transformed data.
+#' @param domain the allowed range of the data to be transformed. The function
+#'   in the `transform` argument is expected to be able to transform the `domain`
+#'   argument.
 #' @seealso \Sexpr[results=rd,stage=build]{scales:::seealso_trans()}
 #' @export
 #' @keywords internal
 #' @aliases trans
-trans_new <- function(name, transform, inverse, breaks = extended_breaks(),
+trans_new <- function(name, transform, inverse,
+                      d_transform = NULL, d_inverse = NULL,
+                      breaks = extended_breaks(),
                       minor_breaks = regular_minor_breaks(),
                       format = format_format(), domain = c(-Inf, Inf)) {
   if (is.character(transform)) transform <- match.fun(transform)
   if (is.character(inverse)) inverse <- match.fun(inverse)
+  if (is.character(d_transform)) d_transform <- match.fun(d_transform)
+  if (is.character(d_inverse)) d_inverse <- match.fun(d_inverse)
 
   structure(
     list(
       name = name,
       transform = transform,
       inverse = inverse,
+      d_transform = d_transform,
+      d_inverse = d_inverse,
       breaks = breaks,
       minor_breaks = minor_breaks,
       format = format,
@@ -93,7 +104,7 @@ as.trans <- function(x, arg = deparse(substitute(x))) {
       compose_trans(!!!x)
     }
   } else {
-    abort(sprintf("`%s` must be a character vector or a transformer object", arg))
+    cli::cli_abort(sprintf("{.arg %s} must be a character vector or a transformer object", arg))
   }
 }