+++ tests/testthat/test-label-currency.R
@@ -1,5 +1,5 @@
 test_that("negative comes before prefix", {
-  expect_equal(label_dollar()(-1), "-$1")
+  expect_equal(label_currency()(-1), "-$1")
 })
 
 test_that("negative_parens is deprecated", {
@@ -10,21 +10,21 @@ test_that("negative_parens is deprecated", {
 })
 
 test_that("preserves NAs", {
-  expect_equal(label_dollar()(NA_real_), NA_character_)
+  expect_equal(label_currency()(NA_real_), NA_character_)
 })
 
 test_that("preserves names", {
-  expect_named(label_dollar()(c(a = 1)), "a")
+  expect_named(label_currency()(c(a = 1)), "a")
 })
 
 test_that("decimal.mark could be modified", {
-  expect_equal(label_dollar(decimal.mark = ",")(123.45), "$123,45")
+  expect_equal(label_currency(decimal.mark = ",")(123.45), "$123,45")
 })
 
 test_that("can rescale with scale_cut", {
-  lab <- label_dollar(scale_cut = cut_short_scale())
+  lab <- label_currency(scale_cut = cut_short_scale())
   expect_equal(lab(c(1, 1e3, 1e6)), c("$1", "$1K", "$1M"))
 
-  lab <- label_dollar(scale_cut = cut_short_scale(), prefix = "", suffix = " USD")
+  lab <- label_currency(scale_cut = cut_short_scale(), prefix = "", suffix = " USD")
   expect_equal(lab(c(1, 1e3, 1e6)), c("1 USD", "1K USD", "1M USD"))
 })